When debootstrapping a Debian Stretch hppa chroot with a qemu-hppa-static it looks sometimes as it hangs or timeouts. Also networking seems not possible.
Running inside a Debian Stretch hppa chroot following
command "chroot stretch su -l root" seems to hang and
fails later. This seems also to be an issue when
debootstrap creates the chroot.

Running with QEMU_STRACE shows that a socket is opened,
something gets sent to it and then some answer is expected:
    4349 getuid(128024,0,0,0,-161706028,-160423560) = 0
    4349 socket(16,SOCK_RAW,9) = 3
    4349 fcntl64(3,F_SETFD,1) = 0
    4349 readlink("/proc/self/exe",0xf6801a48,4096) = 7
    4349 sendto(3,-159367692,108,0,-159358704,12) = 108
    4349 poll(-159358692,1,500,0,-160223744,-159659232) = 0
    4349 recvfrom(3,-159376684,8988,66,-159358520,-159358508) = -1 errno=11 (Resource temporarily unavailable)
      (and repeats poll and recvfrom ...)

Debugging shows that the sending happens in following stack:
  sendto(...)
  audit_send  (libaudit:netlink.c)
  audit_send_user_message  (libaudit:deprecated.c)
  audit_log_acct_message  (libaudit:audit_logging.c)
  _pam_audit_writelog  (libpam:pam_audit.c)
  _pam_auditlog  (libpam:pam_audit.c)
  pam_authenticate  (libpam:pam_auth.c)
  check_perms_pam  (shadow:su.c)
  check_perms  (shadow:su.c)
  main (su.c)

The socket was opened a little bit earlier in
  socket(PF_NETLINK, SOCK_RAW, NETLINK_AUDIT)
  audit_open  (libaudit:netlink.c)
  _pam_audit_open  (libpam:pam_audit.c)
  _pam_auditlog  (libpam:pam_audit.c)
  ...

But the problem is that the message sent in audit_send gets translated
into an empty buffer in do_sendto, because fd_trans_target_to_host_data
has no entry for this fd stored:
    #0  fd_trans_target_to_host_data  (syscall.c:516)
    #1  do_sendto  (syscall.c:3867)
    #2  do_syscall  (syscall.c:9352)
    #3  cpu_loop  (main.c:3607)
    #4  main  (main.c:4726)

When the stocket got opened it did initially call fd_trans_register:
    #0  fd_trans_register  (syscall.c:542)
    #1  do_socket(domain=16, type=3, protocol=9)  (syscall.c:3520)
    #2  do_syscall  (syscall.c:9373)
    #3  cpu_loop  (main.c:3607)
    #4  main  (main.c:4726)

But it gets unregistered again, just a line after the call to do_socket:
    #0  fd_trans_unregister  (syscall.c:557)
    #1  do_syscall  (syscall.c:9375)
    #2  cpu_loop  (main.c:3607)
    #3  main  (main.c:4726)

When commenting this line, this NETLINK_AUDIT operation and
a "apt-get update" started to work.

The "fd_trans_unregister" seems to be included since 2015:
e36800c91a7 (linux-user: add signalfd/signalfd4 syscalls)

Tried to reproduce it with qemu-x86_64-static on amd64, but could
not see this issue. Up to now I see this just with hppa.

Therefore I do not know if this is just my configuration,
because of some "hppa" changes or if this
fd_trans_unregister is really wrong.

What do you think?