target/arm: Update arm_mmu_idx for VHE
Return the indexes for the EL2&0 regime when the appropriate bits
are set within HCR_EL2.  This happens for initial generation in
arm_mmu_idx, and reconstruction in core_to_arm_mmu_idx.

In order to make this reliable, we also need a bit in TBFLAGS.

Reviewed-by: Alex Bennée <alex.bennee@linaro.org>
Signed-off-by: Richard Henderson <richard.henderson@linaro.org>
---
v5: Consistently check E2H & TGE & ELUsingAArch32(EL2).