linux-user: Add PAGE_SHARED
This new bit is set for mmap regions that used MAP_SHARED, or
for shmat regions.  We will use it to notice linux-user uses
of MTE that we cannot support, because we don't have the real
kernel's view of the physical address space.

At the same time, reorder some of the other PAGE_* bits so
that they're in ascending order.  Tidy some of the comments.

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>